--- src/components/Auth/Textarea/Textarea.module.css
@@ -1,24 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.Textarea {
-  resize: none;
-  width: 522px;
-  height: 171px;
-
-  padding: var(--spacing-sm) var(--spacing-base);
-
-  font-family: var(--font-family-content);
-  font-size: var(--text-size-lg);
-  font-weight: var(--text-weight-bold);
-
-  border: none;
-  border-radius: var(--border-radius-sm);
-  box-shadow: var(--shadow-default);
-}
-.label {
-  font-size: var(--text-size-base);
-  font-weight: 600;
-}
