--- src/components/Auth/Textarea/index.ts
@@ -1,1 +1,0 @@
-export * from './Textarea';
