--- src/components/Auth/index.ts
@@ -1,5 +1,0 @@
-export * from './Form';
-export * from './Input';
-export * from './ValidationErrorMessage';
-export * from './ErrorMessage';
-export * from './Textarea';
