+++ src/components/common/Textarea/Textarea.module.css
@@ -1,22 +1,21 @@
 .container {
   display: flex;
   flex-direction: column;
+  align-items: center;
 }
 .Textarea {
   resize: none;
-  width: 343px;
-  height: 292px;
+  width: 522px;
+  height: 171px;
 
   padding: var(--spacing-sm) var(--spacing-base);
 
-  color: var(--grayscale-darkgray);
-
   font-family: var(--font-family-content);
-  font-size: var(--text-size-base);
-  font-weight: 600;
+  font-size: var(--text-size-lg);
+  font-weight: var(--text-weight-bold);
 
   border: none;
-  border-radius: var(--border-radius-md);
+  border-radius: var(--border-radius-sm);
   box-shadow: var(--shadow-default);
 }
 .label {
