+++ src/components/common/Textarea/Textarea.stories.tsx
@@ -1,13 +1,14 @@
-import { Textarea, TextareaProps } from './Textarea';
+import {
+  Textarea,
+  TextareaProps,
+  ValidationErrorMessage,
+} from '@/components/common';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { useForm } from '@/hooks';
 
 export default {
-  title: 'Components/Common/Textarea',
+  title: 'Components/common/Textarea',
   component: Textarea,
-  args: {
-    labelValue: 'Textarea 컴포넌트입니다',
-    placeholder: 'Textarea 컴포넌트입니다',
-  },
   parameters: {
     docs: {
       description: {
@@ -17,14 +18,43 @@ export default {
   },
 } as ComponentMeta<typeof Textarea>;
 
-const Template: ComponentStory<typeof Textarea> = (args: TextareaProps) => (
-  <Textarea {...args} />
-);
+const Template: ComponentStory<typeof Textarea> = (args: TextareaProps) => {
+  const { errors, touched, getFieldProps } = useForm({
+    initialValues: {
+      missionDescription: '',
+    },
+    validate: (values) => {
+      const errors = {
+        missionDescription: '',
+      };
+      if (values.missionDescription.length === 0) {
+        errors.missionDescription = '임무를 작성해 주세요';
+      }
+
+      return errors;
+    },
+    onSubmit: async (values) => {
+      console.log(values);
+    },
+  });
+  return (
+    <Textarea getFieldProps={getFieldProps} {...args}>
+      <ValidationErrorMessage touched={touched} errors={errors} />
+    </Textarea>
+  );
+};
 
 export const Default = Template.bind({});
+Default.args = {
+  placeholder: '좀 더 자세히 설명해 주세요',
+  name: 'missionDescription',
+  labelText: '미션 상세 설명',
+};
 
 export const HiddenLabel = Template.bind({});
 HiddenLabel.args = {
   placeholder: 'label을 숨긴 Textarea 컴포넌트입니다',
+  name: 'missionDescription',
+  labelText: '미션 상세 설명',
   hiddenLabel: true,
 };
